# Improve performance of sequence length computation