--- a/include/utils.h
+++ b/include/utils.h
@@ -296,20 +296,20 @@ namespace utils {
      \return Reversed string of toReverse.
      */
      std::string getReverse(const std::string &toReverse);
+    
+    /**
+     \brief Count the number of occurences of a determined char in the string.
+     \param c Character to count.
+     \param line String to count c in.
+     \return Number of occurences of c character.
+     */
+    int countCharacter(char c, const std::string &line);
 
     /**
      \brief Removes a determined char from the string
      \param c Character to remove from line
      \param line String to remove c from.
      \return New string without c character
-     */
-    int countCharacter(char c, const std::string &line);
-
-    /**
-     \brief Count the number of occurences of a determined char in the string
-     \param c Character to count
-     \param line String to count c in.
-     \return Number of occurences of c character
      */
      std::string removeCharacter(char c, std::string line);