Hi again,

I noticed while working on #72 that there were a couple of locations where a sequence would be copied for the sole purpose of computing its (ungapped) length, which is not really efficient. I added a new dedicated `Alignment` method for getting the length of a sequence without copy, based on the new `utils::countCharacter` function added here as well.

The implementation of `utils::countCharacter` uses [`memchr`](https://cplusplus.com/reference/cstring/memchr/) rather than [`std::count`](https://cplusplus.com/reference/algorithm/count/) or a naive implementation because [it is faster](https://gms.tf/stdfind-and-memchr-optimizations.html). It does increase performance when running in `-clusters` mode (I got a x1.3 to x1.5 improvement):

<details>

<pre>
Benchmark 1: ./trimal_2.0 -in ../dataset/example.015.AA.bctoNOG.ENOG41099F3.fasta -clusters 5
  Time (mean ± σ):     30.104 s ±  1.319 s    [User: 29.951 s, System: 0.051 s]
  Range (min … max):   29.006 s … 31.567 s    3 runs
 
Benchmark 2: ./trimal_memchr -in ../dataset/example.015.AA.bctoNOG.ENOG41099F3.fasta -clusters 5
  Time (mean ± σ):     22.973 s ±  0.485 s    [User: 22.798 s, System: 0.071 s]
  Range (min … max):   22.620 s … 23.526 s    3 runs
 
Summary
  './trimal_memchr -in ../dataset/example.015.AA.bctoNOG.ENOG41099F3.fasta -clusters 5' ran
    1.31 ± 0.06 times faster than './trimal_2.0 -in ../dataset/example.015.AA.bctoNOG.ENOG41099F3.fasta -clusters 5'
</pre>

</details>